+++ scil/code/algo/algo-memcopy.h
@@ -27,7 +27,7 @@
  * \param source_size Byte size of uncompressed buffer
  * \return Success state of the compression
  */
-int scil_memcopy_compress(const scil_context* ctx, byte* restrict dest, uint64_t * restrict out_size, const byte*restrict source, const size_t source_size);
+int scil_memcopy_compress(const scil_context* ctx, byte* restrict dest, size_t * restrict out_size, const byte*restrict source, const size_t source_size);
 
 /**
  * \brief Trivial memcopy decompression function
@@ -38,7 +38,7 @@ int scil_memcopy_compress(const scil_context* ctx, byte* restrict dest, uint64_t
  * \param source_size Byte size of uncompressed buffer
  * \return Success state of the compression
  */
-int scil_memcopy_decompress(const scil_context* ctx, byte*restrict dest, uint64_t exp_size, const byte*restrict source, const size_t source_size);
+int scil_memcopy_decompress(const scil_context* ctx, byte*restrict dest, size_t exp_size, const byte*restrict source, const size_t source_size);
 
 extern scil_compression_algorithm algo_memcopy;
